improvements

+++ nvm-kit.sh
@@ -4,20 +4,22 @@
 MODE="install"
 
 if [[ -z "$1" ]]; then
-    if [ $1 == "uninstall" || $1 == "remove" || $1 == "purge" ]; then
-        MODE="uninstall";
+    MODE="install"
+else
+    if [[ $1 == "uninstall" ]] || [[ $1 == "remove" ]] || [[ $1 == "purge" ]]; then
+        MODE="uninstall"
     fi
 fi
 
 SHELL_DEPENDENCIES="$HOME/.development-tools/.dependencies/shell"
 
-NVM_PATH='# nvm PATH Setup
-export NVM_DIR="$([ -z "${XDG_CONFIG_HOME-}" ] && printf %s "${HOME}/.nvm" || printf %s "${XDG_CONFIG_HOME}/nvm")"
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh" # This loads nvm'
+NVM_PATH='# nvm PATH Setup\nexport NVM_DIR="$([ -z "${XDG_CONFIG_HOME-}" ] && printf %s "${HOME}/.nvm" || printf %s "${XDG_CONFIG_HOME}/nvm")"\n[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh" # This loads nvm\n[ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion\n'
 
-if [[$MODE == "install"]]; then
+if [[ $MODE == "install" ]]; then
+    if [[ ! -e $NVM_DIR ]]; then
+        echo -e $NVM_PATH >> "$SHELL_DEPENDENCIES/paths"
+    fi
     sh -c "$(curl -fsSL https://raw.githubusercontent.com/nvm-sh/nvm/v0.35.3/install.sh)"
-    echo $NVM_PATH >> "$SHELL_DEPENDENCIES/paths"
 
     source "$SHELL_DEPENDENCIES/main"
 
@@ -27,7 +29,11 @@ if [[$MODE == "install"]]; then
     npm i -g typescript
 
 else
-    rm -rf "$NVM_DIR"
-    echo "Please remove the lines below in $SHELL_DEPENDENCIES/paths"
-    echo $NVM_PATH
+    if [[ -e $NVM_DIR ]]; then
+        rm -rf "$NVM_DIR"
+        echo "Please remove the lines below in $SHELL_DEPENDENCIES/paths"
+        echo $NVM_PATH
+    else
+        echo "nvm is not installed"
+    fi
 fi
